requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
torch==2.0.0
torchvision==0.15.1
gunicorn==21.2.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9